Move navMenu into tableView.js where its context is more appropriate
